+++ src/application/application.js
@@ -13,8 +13,7 @@ import { CANVAS, WEBGL, AUTO } from "../const.js";
 
 /**
  * @classdesc
- * An Application represents a single melonJS game.
- * An Application is responsible for updating (each frame) all the related object status and draw them.
+ * An Application represents a single melonJS game, and is responsible for updating (each frame) all the related object status and draw them.
  * @see game
  */
 export default class Application {
@@ -23,10 +22,6 @@ export default class Application {
      * @param {number} height - The height of the canvas viewport
      * @param {Application.Settings} [options] - The optional parameters for the application and default renderer
      * @throws Will throw an exception if it fails to instantiate a renderer
-     * @example
-     * let my game = new Application(640, 480, {renderer: me.video.AUTO}) {
-     *     ....
-     * }
      */
     constructor(width, height, options) {