+++ src/index.js
@@ -203,7 +203,7 @@ export let initialized = false;
  */
 export let skipAutoInit = false;
 
-/*
+/**
  * game is a default instance of a melonJS Application and represents your current game,
  * it contains all the objects, tilemap layers, current viewport, collision map, etc...<br>
  * @namespace game