+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [15.8.0] (melonJS 2) - _2023-xx-xx_
 
 ### Added
-- Physic: new application/world `physic` prorperty allowing to disable the builtin physic implementation
+- Physic: new `physic` flag allowing to disable the builtin physic implementation
 
 ## [15.7.0] (melonJS 2) - _2023-07-19_